rescope arange_patch to patches
